Consolidate all compile-*-expr.js tasks
