--- build-system/tasks/compile-css-expr.js
@@ -1,49 +1,0 @@
-/**
- * Copyright 2016 The AMP HTML Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS-IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-'use strict';
-
-const fs = require('fs-extra');
-const gulp = require('gulp');
-const jison = require('jison');
-
-gulp.task('compile-css-expr', function() {
-  const path = 'extensions/amp-animation/0.1/';
-
-  const bnf = fs.readFileSync(path + 'css-expr-impl.jison', 'utf8');
-  const settings = {
-    type: 'lalr',
-    debug: false,
-    moduleType: 'js',
-    'token-stack': true,
-  };
-  const generator = new jison.Generator(bnf, settings);
-  const jsModule = generator.generate(settings);
-
-  const license = fs.readFileSync(
-      'build-system/tasks/js-license.txt', 'utf8');
-  const suppressCheckTypes = '/** @fileoverview ' +
-      '@suppress {checkTypes, suspiciousCode, uselessCode} */';
-  const imports = 'import * as ast from \'./css-expr-ast\';';
-  const jsExports = 'export const cssParser = parser;';
-
-  const out = [
-    license,
-    suppressCheckTypes,
-    imports,
-    jsModule,
-    jsExports].join('\n\n') + '\n';
-  fs.writeFileSync(path + 'css-expr-impl.js', out);
-});
