+++ build-system/tasks/index.js
@@ -21,9 +21,7 @@ require('./changelog');
 require('./check-links');
 require('./clean');
 require('./compile');
-require('./compile-access-expr');
-require('./compile-bind-expr');
-require('./compile-css-expr');
+require('./compile-expr');
 require('./create-golden-css');
 require('./csvify-size');
 require('./dep-check');
