+++ extensions/amp-access/0.1/access-expr-impl.js
@@ -17,6 +17,8 @@
 
 /** @fileoverview @suppress {checkTypes, suspiciousCode, uselessCode} */
 
+
+
 /* parser generated by jison 0.4.18 */
 /*
   Returns a Parser object of the following structure:
